refactor: 바우처 클래스 리팩토링
